Refactorisation du code.

+++ api/src/main/java/com/discoodle/api/service/RoomService.java
@@ -24,19 +24,20 @@ public class RoomService {
         );
         Room finalRoom = roomRepository.save(room);
         for (Long room_member : room_members) {
-            Optional<User> temp = userRepository.findById(room_member);
-            if (temp.isPresent()) {
+            if (userRepository.existsById(room_member)) {
                 roomRepository.addNewMember(finalRoom.getRoom_id(),room_member);
-                finalRoom.getUsers().add(temp.get());
             }
         }
         return finalRoom;
     }
 
     public Optional<Room> removeMember(String room_id, Long user_id) {
-        Optional<Room> room = roomRepository.findRoomByUUID(room_id);
-        roomRepository.removeMember(room_id, user_id);
-        return room;
+        if(roomRepository.existsById(room_id)) {
+            Optional<Room> room = roomRepository.findRoomByUUID(room_id);
+            roomRepository.removeMember(room_id, user_id);
+            return room;
+        }
+        return Optional.empty();
     }
 
     public void addNewMember(String room_id, Long user_id) {
